requirements.txt
streamlit>=1.33
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.1.7
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
openai>=1.40.0
tiktoken>=0.7.0
python-dotenv>=1.0.1

